# docs/openapi/season-service-v1.yaml
openapi: 3.1.0
info:
  title: ACM Season Service API
  description: API documentation for ACM Season Service
  license:
    name: Apache 2.0
    url: https://springdoc.org
  version: 1.0.0
servers:
- url: http://localhost
  description: Generated server url
security:
- bearerAuth: []
paths:
  /api/v1/workspace/tasks/{id}:
    get:
      tags:
      - task-workspace-controller
      summary: Get task
      description: Get task details by ID
      operationId: getTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
    put:
      tags:
      - task-workspace-controller
      summary: Update task
      description: Update task details
      operationId: updateTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
    delete:
      tags:
      - task-workspace-controller
      summary: Delete task
      description: Permanently delete task (dev only)
      operationId: deleteTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/v1/tasks/{id}:
    get:
      tags:
      - task-controller
      summary: Get task detail
      description: Get task detail if its season belongs to current farmer
      operationId: getTask_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Task not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
    put:
      tags:
      - task-controller
      summary: Update task
      description: Update task details while task and season are still editable
      operationId: updateTask_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Task not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
    delete:
      tags:
      - task-controller
      summary: Delete task
      description: Delete task if allowed by status and season constraints
      operationId: deleteTask_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "404":
          description: Task not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/v1/seasons/{id}:
    get:
      tags:
      - season-controller
      summary: Get season detail
      description: Get season detail belonging to current farmer
      operationId: getSeason
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
    put:
      tags:
      - season-controller
      summary: Update season
      description: Update season details while status is still editable
      operationId: updateSeason
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSeasonRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
    delete:
      tags:
      - season-controller
      summary: Delete season
      description: Delete a season if allowed by business constraints
      operationId: deleteSeason
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /api/v1/workspace/tasks:
    get:
      tags:
      - task-workspace-controller
      summary: List tasks
      description: List all tasks for current user with filters
      operationId: listTasks
      parameters:
      - name: status
        in: query
        description: Filter by status
        required: false
        schema:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - DONE
          - OVERDUE
          - CANCELLED
      - name: seasonId
        in: query
        description: Filter by season ID
        required: false
        schema:
          type: integer
          format: int32
      - name: q
        in: query
        description: Search by title (min 2 chars)
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: Page number (0-based)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sortBy
        in: query
        description: Sort by field
        required: false
        schema:
          type: string
          default: createdAt
      - name: sortDirection
        in: query
        description: Sort direction
        required: false
        schema:
          type: string
          default: desc
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
    post:
      tags:
      - task-workspace-controller
      summary: Create task
      description: Create a new task with optional season link
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
  /api/v1/seasons:
    get:
      tags:
      - season-controller
      summary: Search seasons
      description: "Search seasons of the current farmer by plot, crop, status and\
        \ date range"
      operationId: searchSeasons
      parameters:
      - name: plotId
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: cropId
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        description: From date (yyyy-MM-dd)
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        description: To date (yyyy-MM-dd)
        required: false
        schema:
          type: string
          format: date
      - name: page
        in: query
        description: Page index (0-based)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseSeasonResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseSeasonResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseSeasonResponse"
    post:
      tags:
      - season-controller
      summary: Create season
      description: Create a new season for a given plot and crop
      operationId: createSeason
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeasonRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonDetailResponse"
  /api/v1/seasons/{seasonId}/tasks:
    get:
      tags:
      - task-controller
      summary: List tasks of a season
      description: List tasks for a given season of the current farmer
      operationId: listTasks_1
      parameters:
      - name: seasonId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        description: From date (yyyy-MM-dd)
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        description: To date (yyyy-MM-dd)
        required: false
        schema:
          type: string
          format: date
      - name: page
        in: query
        description: Page index (0-based)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
        "404":
          description: Season not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
    post:
      tags:
      - task-controller
      summary: Create task in season
      description: Create a new task within given season for current farmer
      operationId: createTask_1
      parameters:
      - name: seasonId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Season not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
...
  /api/v1/seasons/{id}/start:
    post:
      tags:
      - season-controller
      summary: Start season
      description: Start a planned season (PLANNED → ACTIVE)
      operationId: startSeason
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartSeasonRequest"
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonResponse"
        "400":
          description: Invalid status transition
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonResponse"
        "404":
          description: Season not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonResponse"
...
  /api/v1/seasons/{id}/cancel:
    post:
      tags:
      - season-controller
      summary: Cancel season
      description: Cancel a season (→ CANCELLED)
      operationId: cancelSeason
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelSeasonRequest"
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonResponse"
        "400":
          description: Invalid status transition or harvests exist
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonResponse"
        "404":
          description: Season not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeasonResponse"
...
  /api/v1/workspace/tasks/{id}/start:
    patch:
      tags:
      - task-workspace-controller
      summary: Start task
      description: Mark task as started (IN_PROGRESS)
      operationId: startTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartTaskRequest"
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
  /api/v1/workspace/tasks/{id}/done:
    patch:
      tags:
      - task-workspace-controller
      summary: Mark task as done
      description: Mark task as completed (DONE)
      operationId: doneTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskDoneRequest"
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
...
components:
  schemas:
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
          maxLength: 4000
          minLength: 0
        seasonId:
          type: integer
          format: int32
        plannedDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        notes:
          type: string
          maxLength: 4000
          minLength: 0
        assigneeUserId:
          type: integer
          format: int64
      required:
      - title
    ApiResponseTaskResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          $ref: "#/components/schemas/TaskResponse"
        data:
          $ref: "#/components/schemas/TaskResponse"
        generatedAt:
          type: string
...
    StartSeasonRequest:
      type: object
      properties:
        actualStartDate:
          type: string
          format: date
...
    CancelSeasonRequest:
      type: object
      properties:
        forceCancel:
          type: boolean
        reason:
          type: string
...
    ApiResponseVoid:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          type: object
        data:
          type: object
        generatedAt:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT